requirements.txt
# Framework
fastapi==0.116.1
uvicorn==0.35.0
pydantic[email]==2.11.7

# Database
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

# Encrypt / Auth
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# AWS
boto3==1.40.17

# Utils 
python-dotenv==1.1.1
httpx==0.28.1

# Unittest
pytest==8.4.1
pytest-cov==6.2.1
